Split AddressCandidateHelper into expect and actual implementation
